Add users page.
Also add JSP support & version of users page in JSP. It seems to load in roughly the same time, but only for pages with no external requests. This is because it needs to create the HTML internally first. Things like fonts and externa CSS would increase JSP's load time compared to HTML.